fix: merge

+++ internal/core/runner/python/setup.go
@@ -26,28 +26,42 @@ const (
 )
 
 func init() {
-	releaseLibBinary()
+	releaseLibBinary(true)
 }
 
-func releaseLibBinary() {
+func releaseLibBinary(force_remove_old_lib bool) {
 	log.Info("initializing python runner environment...")
 	// remove the old lib
 	if _, err := os.Stat(path.Join(LIB_PATH, LIB_NAME)); err == nil {
-		err := os.Remove(path.Join(LIB_PATH, LIB_NAME))
-		if err != nil {
-			log.Panic(fmt.Sprintf("failed to remove %s", path.Join(LIB_PATH, LIB_NAME)))
-		}
-	}
+		if force_remove_old_lib {
+			err := os.Remove(path.Join(LIB_PATH, LIB_NAME))
+			if err != nil {
+				log.Panic(fmt.Sprintf("failed to remove %s", path.Join(LIB_PATH, LIB_NAME)))
+			}
 
-	err := os.MkdirAll(LIB_PATH, 0755)
-	if err != nil {
-		log.Panic(fmt.Sprintf("failed to create %s", LIB_PATH))
+			// write the new lib
+			err = os.MkdirAll(LIB_PATH, 0755)
+			if err != nil {
+				log.Panic(fmt.Sprintf("failed to create %s", LIB_PATH))
+			}
+			err = os.WriteFile(path.Join(LIB_PATH, LIB_NAME), python_lib, 0755)
+			if err != nil {
+				log.Panic(fmt.Sprintf("failed to write %s", path.Join(LIB_PATH, LIB_NAME)))
+			}
+		}
+	} else if err == os.ErrNotExist {
+		err = os.MkdirAll(LIB_PATH, 0755)
+		if err != nil {
+			log.Panic(fmt.Sprintf("failed to create %s", LIB_PATH))
+		}
+		err = os.WriteFile(path.Join(LIB_PATH, LIB_NAME), python_lib, 0755)
+		if err != nil {
+			log.Panic(fmt.Sprintf("failed to write %s", path.Join(LIB_PATH, LIB_NAME)))
+		}
+		log.Info("python runner environment initialized")
+	} else {
+		log.Panic(fmt.Sprintf("failed to initialize python runner environment: %v", err))
 	}
-	err = os.WriteFile(path.Join(LIB_PATH, LIB_NAME), python_lib, 0755)
-	if err != nil {
-		log.Panic(fmt.Sprintf("failed to write %s", path.Join(LIB_PATH, LIB_NAME)))
-	}
-	log.Info("python runner environment initialized")
 }
 
 func checkLibAvaliable() bool {
